tweak tests

+++ .ci/run_tests.sh
@@ -13,20 +13,20 @@ set -e
 (cd tests-native; make -B CONTEXT=valgrind EXTRA_COMPILE_ARGS='-D_GLIBCXX_DEBUG -D_GLIBCXX_DEBUG_PEDANTIC' test)
 (cd tests-native; make -B CXX=clang++-8 CC=clang-8 OPTIMIZE=1 WITH_OPENMP=0 EXTRA_COMPILE_ARGS='-fsanitize=address -DNDEBUG' test)
 
-rm -r build/
+if [ -e build/ ]; then >&2 echo "Old build dir present?"; exit 1; fi
 CFLAGS="-D_GLIBCXX_DEBUG -D_GLIBCXX_DEBUG_PEDANTIC" python3 setup.py build_ext -i
 bash -c "ulimit -v 2048000; ./scripts/run_tests.sh"
 
-rm -r build/
+rm -rf build/
 CC=clang-8 \
   CXX=clang++-8 \
   CFLAGS="-fsanitize=address -UNDEBUG" \
   python3 setup.py build_ext -i
-ASAN_OPTIONS=detect_leaks=0 LD_PRELOAD=/usr/lib/llvm-8/lib/clang/8.0.1/lib/linux/libclang_rt.asan-x86_64.so ./scripts/run_tests.sh "${@:2}"
+PYTHONMALLOC=malloc ASAN_OPTIONS=detect_leaks=0 LD_PRELOAD=/usr/lib/llvm-8/lib/clang/8.0.1/lib/linux/libclang_rt.asan-x86_64.so ./scripts/run_tests.sh "${@:2}"
 
 python3 -m pip uninstall -y chemreac
 
-rm -r build/
+rm -rf build/
 
 python3 setup.py sdist
 cp dist/${PKG_NAME}-*.tar.gz /tmp
